style: 플레이스 페이지 UI 레이아웃 수정 (#91)
